+++ src/thor/route_matcher.cc
@@ -1,3 +1,4 @@
+#include "baldr/tilehierarchy.h"
 #include "midgard/logging.h"
 #include "midgard/util.h"
 #include <algorithm>
@@ -24,6 +25,12 @@ namespace {
 
 using end_edge_t = std::pair<odin::Location::PathEdge, float>;
 using end_node_t = std::unordered_map<GraphId, end_edge_t>;
+
+// Data type to record edges and transitions that have been followed for each shape
+// index and hierarchy level. First entry in the pair is the index (from the node's
+// starting directed edge index) of the directed edge. The second entry in the pair
+// is the transition index.
+using followed_edges_t = std::vector<std::vector<std::pair<uint32_t, uint32_t>>>;
 
 // Total distance match delta
 constexpr float kTotalDistanceEpsilon = 5.0f;
@@ -86,11 +93,15 @@ end_node_t GetEndEdges(GraphReader& reader,
   return end_nodes;
 }
 
+// Expand from a correlated node. Walks the shape ahead to find the next correlated
+// node and expands from there. Returns true once the end node has been found (and
+// distance is approximately what it should be). Returns false if expansion from this
+// node fails (cannot find edges that match the trace - either in position or distance).
 bool expand_from_node(const std::shared_ptr<DynamicCost>* mode_costing,
                       const TravelMode& mode,
                       GraphReader& reader,
                       const std::vector<meili::Measurement>& shape,
-                      const std::vector<float>& distances,
+                      std::vector<std::pair<float, float>>& distances,
                       const bool use_timestamps,
                       size_t& correlated_index,
                       const GraphTile* tile,
@@ -101,29 +112,29 @@ bool expand_from_node(const std::shared_ptr<DynamicCost>* mode_costing,
                       std::vector<PathInfo>& path_infos,
                       const bool from_transition,
                       GraphId& end_node,
-                      const float total_distance) {
-  // Done expanding when node equals stop node and the accumulated distance plus
-  // the partial last edge distance is approximately equal to the total distance
+                      followed_edges_t& followed_edges) {
+  // Done expanding when node equals stop node and the accumulated distance to that node
+  // plus the partial last edge distance is approximately equal to the total distance
   auto n = end_nodes.find(node);
-  if (n != end_nodes.end()) {
-    // Accumulate distance through correlated index
-    float acc_distance = 0.0f;
-    for (uint32_t idx = 0; idx <= correlated_index; idx++) {
-      acc_distance += distances[idx];
-    }
-
-    // Verify distance is approximately equal to the total distance
-    if (midgard::equal<float>((acc_distance + n->second.second), total_distance,
-                              kTotalDistanceEpsilon)) {
-      end_node = node;
-      return true;
-    }
+  if (n != end_nodes.end() &&
+      midgard::equal<float>((distances[correlated_index].second + n->second.second),
+                            distances.back().second, kTotalDistanceEpsilon)) {
+    end_node = node;
+    return true;
   }
 
+  // Get the last edge followed from this index
+  uint32_t level = node.level();
+  uint32_t start_de = followed_edges[correlated_index][level].first;
+
+  // Iterate through directed edges from this node
   const NodeInfo* node_info = tile->node(node);
-  GraphId edge_id(node.tileid(), node.level(), node_info->edge_index());
+  GraphId edge_id(node.tileid(), level, node_info->edge_index());
   const DirectedEdge* de = tile->directededge(node_info->edge_index());
-  for (uint32_t i = 0; i < node_info->edge_count(); i++, de++, ++edge_id) {
+  for (uint32_t i = start_de; i < node_info->edge_count(); i++, de++, ++edge_id) {
+    // Mark the directed edge as already followed
+    followed_edges[correlated_index][level].first = i;
+
     // Skip shortcuts and transit connection edges
     // TODO - later might allow transit connections for multi-modal
     if (de->is_shortcut() || de->use() == Use::kTransitConnection) {
@@ -145,14 +156,13 @@ bool expand_from_node(const std::shared_ptr<DynamicCost>* mode_costing,
     PointLL de_end_ll = end_node_tile->get_node_ll(de->endnode());
     float de_length = length_comparison(de->length(), true);
 
-    // Process current edge until shape matches end node
-    // or shape length is longer than the current edge. Increment to the
-    // next shape point after the correlated index.
+    // Process current edge until shape matches end node or shape length is longer than
+    // the current edge. Increment to the next shape point after the correlated index.
     size_t index = correlated_index + 1;
     float length = 0.0f;
     while (index < shape.size()) {
       // Exclude edge if length along shape is longer than the edge length
-      length += distances.at(index);
+      length += distances.at(index).first;
       if (length > de_length) {
         break;
       }
@@ -178,11 +188,10 @@ bool expand_from_node(const std::shared_ptr<DynamicCost>* mode_costing,
         // Continue walking shape to find the end edge...
         if (expand_from_node(mode_costing, mode, reader, shape, distances, use_timestamps, index,
                              end_node_tile, de->endnode(), end_nodes, prev_edge_label, elapsed_time,
-                             path_infos, false, end_node, total_distance)) {
+                             path_infos, false, end_node, followed_edges)) {
           return true;
         } else {
-          // Match failed along this edge, pop the last entry off path_infos
-          // and try the next edge
+          // Match failed along this edge, pop the last entry off path_infos and try the next edge
           path_infos.pop_back();
           break;
         }
@@ -191,10 +200,14 @@ bool expand_from_node(const std::shared_ptr<DynamicCost>* mode_costing,
     }
   }
 
-  // Handle transitions - expand from the end node of each transition
+  // Get the last transition followed from this index
+  uint32_t start_trans = followed_edges[correlated_index][level].second;
+
+  // Handle transitions - expand from the transition end nodes
   if (!from_transition && node_info->transition_count() > 0) {
     const NodeTransition* trans = tile->transition(node_info->transition_index());
-    for (uint32_t i = 0; i < node_info->transition_count(); ++i, ++trans) {
+    for (uint32_t i = start_trans; i < node_info->transition_count(); ++i, ++trans) {
+      followed_edges[correlated_index][level].second = i;
       const GraphTile* end_node_tile = reader.GetGraphTile(trans->endnode());
       if (end_node_tile == nullptr) {
         continue;
@@ -202,7 +215,7 @@ bool expand_from_node(const std::shared_ptr<DynamicCost>* mode_costing,
       if (expand_from_node(mode_costing, mode, reader, shape, distances, use_timestamps,
                            correlated_index, end_node_tile, trans->endnode(), end_nodes,
                            prev_edge_label, elapsed_time, path_infos, true, end_node,
-                           total_distance)) {
+                           followed_edges)) {
         return true;
       }
     }
@@ -220,14 +233,26 @@ bool RouteMatcher::FormPath(const std::shared_ptr<DynamicCost>* mode_costing,
                             const bool use_timestamps,
                             const google::protobuf::RepeatedPtrField<odin::Location>& correlated,
                             std::vector<PathInfo>& path_infos) {
-  // Form distances between shape points
+  if (shape.size() < 2) {
+    throw std::runtime_error("Invalid shape - less than 2 points");
+  }
+
+  // Form distances between shape points and accumulated distance from start to each shape point
   float total_distance = 0.0f;
-  std::vector<float> distances;
-  distances.push_back(0.0f);
+  std::vector<std::pair<float, float>> distances;
+  distances.push_back(std::make_pair(0.0f, 0.0f));
   for (size_t i = 1; i < shape.size(); i++) {
     float d = shape[i].lnglat().Distance(shape[i - 1].lnglat());
-    distances.push_back(d);
     total_distance += d;
+    distances.push_back(std::make_pair(d, total_distance));
+  }
+
+  // Keep a record of followed edges and transition from each shape index (for each hierarchy level) -
+  // this prevents doubling back and causing an infinite loop (could be due to transitions)
+  followed_edges_t followed_edges;
+  followed_edges.resize(shape.size());
+  for (auto& f : followed_edges) {
+    f.resize(TileHierarchy::get_max_level());
   }
 
   // Process and validate end edges (can be more than 1). Create a map of
@@ -266,9 +291,10 @@ bool RouteMatcher::FormPath(const std::shared_ptr<DynamicCost>* mode_costing,
     float de_remaining_length = de->length() * (1 - edge.percent_along());
     float de_length = length_comparison(de_remaining_length, true);
     EdgeLabel prev_edge_label;
+
     // Loop over shape to form path from matching edges
     while (index < shape.size()) {
-      length += distances.at(index);
+      length += distances.at(index).first;
       if (length > de_length) {
         break;
       }
@@ -295,9 +321,8 @@ bool RouteMatcher::FormPath(const std::shared_ptr<DynamicCost>* mode_costing,
         GraphId end_node;
         if (expand_from_node(mode_costing, mode, reader, shape, distances, use_timestamps, index,
                              end_node_tile, de->endnode(), end_nodes, prev_edge_label, elapsed_time,
-                             path_infos, false, end_node, total_distance)) {
-          // If node equals stop node then when are done expanding - get
-          // the matching end edge
+                             path_infos, false, end_node, followed_edges)) {
+          // If node equals stop node then when are done expanding - get the matching end edge
           auto n = end_nodes.find(end_node);
           if (n == end_nodes.end()) {
             return false;
@@ -332,10 +357,9 @@ bool RouteMatcher::FormPath(const std::shared_ptr<DynamicCost>* mode_costing,
 
           // Add end edge
           path_infos.emplace_back(mode, elapsed_time, end_edge_graphid, 0);
-
           return true;
         } else {
-          // Did not find end edge - so get out
+          // Did not find an edge that correlates with the trace, return false.
           return false;
         }
       }